requirements.txt
ollama
mistralai
google-generativeai